requirements.txt
fastapi
boto3
pydantic
uvicorn
PyJWT[crypto]==2.3.0 OR PyJWT == 2.3.0
python-multipart
python-dotenv


AWS_ACCESS_KEY_ID=AKIA4MI2JP2FOQXW4QGJ
AWS_SECRET_ACCESS_KEY=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
AWS_REGION=us-east-1
COGNITO_USER_POOL_CLIENT_ID=33jhkmuko5a7r4flp8q28dk1ad
COGNITO_CLIENT_SECRET=ml0ds53dlsn0ardc4d5oj2mqtsk87n76tqgdqhemp3e6a41k0ha

